requirements.txt
pillow==9.4.0
numpy==1.24.1
scipy==1.10.0
humanize==4.4.0
aioredis==2.0.1
hiredis==2.1.1

pycairo==1.23.0
PyGObject==3.42.2

ipython>=8.8.0
flake8>=6.0.0
black>=22.12.0
PyGObject-stubs>=2.2.0